Create notes directory and rename router to index
